feat(toast): add toastsFeature flag; fix HMR and provider ordering
- Add toastsFeature flag (disabled: []) to featureFlags.ts and FeatureFlagKey
- Gate ToastManager rendering behind useFeatureFlag('toastsFeature')
- Move useToast hook to src/hooks/useToast.ts so ToastContext only exports
  components (fixes Vite Fast Refresh HMR warning)
- Move ToastProvider inside GameProvider in App.tsx so useFeatureFlag →
  useGame resolves correctly (fixes runtime crash)
