+++ frontend/src/hooks/useToast.ts
@@ -1,0 +1,11 @@
+import { useContext } from 'react';
+import { ToastContext } from '../context/ToastContext';
+import type { ToastContextValue } from '../context/ToastContext';
+
+export function useToast(): ToastContextValue {
+  const context = useContext(ToastContext);
+  if (!context) {
+    throw new Error('useToast must be used within a ToastProvider');
+  }
+  return context;
+}
